+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 * Building Realm-Xcode6 for iOS now builds a universal framework for Simulator & Device.
 * Using NSNumber properties (unsupported) now throws a more informative exception.
+* Added `[RLMRealm defaultRealmPath]`
 
 ### Bugfixes